% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{imf.data-package}
\alias{imf.data}
\alias{imf.data-package}
\title{imf.data: An Interface to IMF (International Monetary Fund) Data JSON API}
\description{
A straightforward interface for accessing the IMF (International Monetary Fund) data JSON API, available at \url{https://data.imf.org/}. This package offers direct access to the primary API endpoints: Dataflow, DataStructure, and CompactData. And, it provides an intuitive interface for exploring available dimensions and attributes, as well as querying individual time-series datasets. Additionally, the package implements a rate limit on API calls to reduce the chances of exceeding service limits (limited to 10 calls every 5 seconds) and encountering response errors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedrobtz.github.io/imf.data/}
  \item Report bugs at \url{https://github.com/pedrobtz/imf.data/issues}
}

}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com}

}
\keyword{internal}
