% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mt_data_structure}
\alias{mt_data_structure}
\title{Calls API endpoint 'DataStructure'}
\usage{
mt_data_structure(id)
}
\arguments{
\item{id}{is a Dataset Series identifier.}
}
\value{
a list with content of 'DataStructure' response.
}
\description{
Calls API endpoint 'DataStructure'
}
\examples{
DOT <- mt_data_structure("DOT")

}
