% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mt_dataflow}
\alias{mt_dataflow}
\title{Calls API endpoint 'Dataflow'}
\usage{
mt_dataflow()
}
\value{
a list with content of 'Dataflow' response.
}
\description{
Calls API endpoint 'Dataflow'
}
\examples{
mt_dataflow()

}
