% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_api.R
\name{sdmx_request}
\alias{sdmx_request}
\title{Create a base request for the IMF SDMX API}
\usage{
sdmx_request(path, format = c("json", "xml"))
}
\arguments{
\item{path}{Character. The API endpoint path.}

\item{format}{Character. Response format: "json" or "xml". Default is "json".}
}
\value{
An httr2 request object.
}
\description{
Create a base request for the IMF SDMX API
}
\keyword{internal}
