% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_datasets}
\alias{list_datasets}
\title{List all available Time Series Datasets}
\usage{
list_datasets()
}
\value{
a data.frame with columns 'Id' and 'Description'.
}
\description{
List all available Time Series Datasets
}
\examples{
d <- list_datasets()
head(d)

}
