% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_api.R
\name{sdmx_metadata}
\alias{sdmx_metadata}
\title{Query metadata from the IMF SDMX API}
\usage{
sdmx_metadata(
  provider_id = "*",
  metadataset_id = "*",
  version = "+",
  detail = c("allstubs", "full"),
  format = c("json", "xml")
)
}
\arguments{
\item{provider_id}{Character. Provider ID. Default is "*".}

\item{metadataset_id}{Character. Metadataset ID. Default is "*".}

\item{version}{Character. Version. Default is "+" (latest stable).}

\item{detail}{Character. Detail level: "full" or "allstubs". Default is "allstubs".}

\item{format}{Character. Response format: "json" or "xml". Default is "json".}
}
\value{
Parsed API response.
}
\description{
Retrieve metadatasets from the API.
}
\examples{
\dontrun{
# Get all metadatasets
sdmx_metadata()
}
}
